+++ python/pounce/jax/__init__.py
@@ -45,6 +45,7 @@ _jax_config.update("jax_enable_x64", True)
 from ._build import from_jax
 from ._diff import solve, solve_with_warm, vmap_solve, vmap_solve_parallel
 from ._problem import AnchorState, JaxProblem
+from ._path import PathFollower, PathTrace
 
 __all__ = [
     "from_jax",
@@ -54,4 +55,6 @@ __all__ = [
     "vmap_solve_parallel",
     "JaxProblem",
     "AnchorState",
+    "PathFollower",
+    "PathTrace",
 ]